+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 
 v0.14.12 (??.02.2015)
  * Fixed absolute url in link util, if the url are already absolute
+ * New isEmpty method in data model
  
 v0.14.11 (03.02.2015) (TER release)
  * New method deleteReferencesByReference to remove FAL references by reference id(s)